Move and_reduction::prove into its own prove.rs
and_reduction/mod.rs mixed module wiring (declarations, re-exports) with the substantial
prove() function and its own test module — over 100 of the file's 155 lines. The sibling
protocol in this exact codebase, protocols/shift/mod.rs, keeps mod.rs as pure wiring and
gives its equivalent prove() its own shift/prove.rs.

Moved prove() and its tests into a new and_reduction/prove.rs, matching that convention.
mod.rs shrinks to 11 lines of declarations and re-exports. The one substantive edit the move
required: `use self::prover::OblongZerocheckProver;` becomes `use super::prover::...;`, since
prove.rs is now a sibling of prover, not its parent.

Pure move: no behavior, signature, or test changes. Public surface unchanged — confirmed
against every external caller (crates/prover/src/prove.rs, crates/m4-prover/src/prove.rs,
crates/m4-prover/src/witness/operand_columns.rs, the and_reduction benchmark).
